# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CryptoStake Casino API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserPage"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get a user by ID (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/kyc:
    get:
      operationId: listAdminKyc
      tags: [admin]
      summary: List KYC submissions pending review
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: KYC submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminKycPage"
  /admin/kyc/{id}:
    patch:
      operationId: reviewKyc
      tags: [admin]
      summary: Approve or reject a KYC submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycReview"
      responses:
        "200":
          description: Updated KYC record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycRecord"
  /admin/transactions:
    get:
      operationId: listAdminTransactions
      tags: [admin]
      summary: List all transactions (admin)
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [deposit, withdrawal, bet, win, loss]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, userId, type, currency, amount, status, createdAt]
      properties:
        id:          { type: integer }
        userId:      { type: integer }
        type:        { type: string, enum: [deposit, withdrawal, bet, win, loss] }
        currency:    { type: string }
        amount:      { type: string }
        status:      { type: string, enum: [pending, confirmed, failed] }
        txHash:      { type: ["string", "null"] }
        toAddress:   { type: ["string", "null"] }
        gameType:    { type: ["string", "null"] }
        createdAt:   { type: string }
    TransactionPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }

    # ─── KYC ─────────────────────────────────────────────────────────────────
    KycRecord:
      type: object
      required: [id, userId, status, createdAt]
      properties:
        id:          { type: integer }
        userId:      { type: integer }
        status:      { type: string, enum: [pending, approved, rejected] }
        fullName:    { type: ["string", "null"] }
        dateOfBirth: { type: ["string", "null"] }
        country:     { type: ["string", "null"] }
        rejectionReason: { type: ["string", "null"] }
        createdAt:   { type: string }
        updatedAt:   { type: string }
...
    KycReview:
      type: object
      required: [status]
      properties:
        status:          { type: string, enum: [approved, rejected] }
        rejectionReason: { type: string }

    # ─── Games ───────────────────────────────────────────────────────────────
...
    AdminUser:
      type: object
      required: [id, clerkId, username, email, createdAt, kycStatus, isAdmin, isBanned]
      properties:
        id:          { type: integer }
        clerkId:     { type: string }
        username:    { type: string }
        email:       { type: string }
        createdAt:   { type: string }
        kycStatus:   { type: string }
        isAdmin:     { type: boolean }
        isBanned:    { type: boolean }
        totalWagers: { type: string }
        totalDeposits: { type: string }
    AdminUserUpdate:
      type: object
      properties:
        isAdmin:  { type: boolean }
        isBanned: { type: boolean }
    AdminUserPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }
    AdminKycPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/KycRecord"
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }
...