# lib/api-zod/src/generated/types/adminKycPage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CryptoStake Casino API
 * OpenAPI spec version: 0.1.0
 */
import type { KycRecord } from './kycRecord';

export interface AdminKycPage {
  items: KycRecord[];
  total: number;
  page: number;
  limit: number;
}